Added deployment cache
